Added a major revision
1. Changed the way the code handles the representation of grids.  Now grids are
specified in chunks of spheres.  There is a primitive atomic grid that is
computed as if the radial parameter is 1.000 and the real atomic grids are
pointed to this with the proper scaling factor.  This also facilitated the
use of pruned grids.

2. Removed the Atomic Size adjustment stuff in the weighting functions as
nobody else uses it.

3. Added the SG-1 grid as specified in the paper Gill, Johnson, and Pople,
Chem. Phys. Lett. 209, pp. 506


git-svn-id: https://sirius.chem.vt.edu/svn/psi3/trunk@212 29f243b5-bcef-41bb-bbe7-8e20262f3206