A number of extensions and bug-fixes:
(1) dot14.c, dot23.c, dot24.c: Error in memory allocation -- specifically,
using block_memory() with dpd_free_block() -- was causing problems with the
DPD memory accounting system.

(2) buf4_sort_axpy.c: Added out-of-core pqsr, prqs, prsq, rpqs, srqp, and
rspq cases.

(3) Added a "start13" integer matrix to the dpdparam4 struct (and thus to
every dpdbuf4 and dpdfile4.  This computes row offsets when one need to
decompose a (pq,rs) buf4/file4 into (p,qrs), including symmetry.  This is
related to the shift13 structure, but requires less overhead and is useful
for out-of-core algorithms, including some in the Wabei terms in cchbar and
in triples codes.

(4) Removed a lot of excess printing associated with out-of-core algs.

-TDC


git-svn-id: https://sirius.chem.vt.edu/svn/psi3/trunk@2345 29f243b5-bcef-41bb-bbe7-8e20262f3206